feat: support scatter axes options
